+++ spec/integration/braintree/customer_gateway_spec.js
@@ -259,8 +259,14 @@ describe('CustomerGateway', function () {
 
       it('creates a customer with a Us Bank Account nonce', done =>
         specHelper.generateValidUsBankAccountNonce(function (token) {
-          let customerParams =
-            {paymentMethodNonce: token};
+          let customerParams = {
+            paymentMethodNonce: token,
+            creditCard: {
+              options: {
+                verificationMerchantAccountId: 'us_bank_merchant_account'
+              }
+            }
+          };
 
           specHelper.defaultGateway.customer.create(customerParams, function (err, response) {
             assert.isNull(err);
@@ -629,7 +635,7 @@ describe('CustomerGateway', function () {
       specHelper.defaultGateway.customer.create(customerParams, function (err, response) {
         assert.isNull(err);
         assert.isTrue(response.success);
-        assert.isTrue(response.customer.creditCards[0].venmoSdk);
+        assert.isFalse(response.customer.creditCards[0].venmoSdk);
 
         done();
       });
@@ -740,8 +746,14 @@ describe('CustomerGateway', function () {
 
     it('returns us bank account for a given customer', done =>
       specHelper.generateValidUsBankAccountNonce(function (token) {
-        let customerParams =
-          {paymentMethodNonce: token};
+        let customerParams = {
+          paymentMethodNonce: token,
+          creditCard: {
+            options: {
+              verificationMerchantAccountId: 'us_bank_merchant_account'
+            }
+          }
+        };
 
         specHelper.defaultGateway.customer.create(customerParams, function (err, response) {
           assert.isNull(err);