+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2.9.0
+* Add `oauthAccessRevocation` to `WebhookNotification`s
+* Add support for dispute search by `customerId`, `effectiveDate`, and `disbursementDate`
+* Remove `sepaMandateType` and `sepaMandateAcceptanceLocation` params from `ClientTokenGateway`
+* Add support for categorized dispute evidence
+
 ## 2.8.0
 * Fix binary multipart uploads.
 * Add support for `association_filter_id` in `Customer#find`